# docs/api/openapi-design.yaml
openapi: 3.1.0
info:
  title: EnergyForecast API
  version: 0.1.0-design
  description: >
    Design-time contract for the EnergyForecast course project.
    The FastAPI-generated OpenAPI document becomes authoritative after implementation.
servers:
  - url: /api/v1
...
paths:
...
  /experiments:
    get:
      tags: [Experiments]
      operationId: listExperiments
      parameters:
        - {$ref: "#/components/parameters/Page"}
        - {$ref: "#/components/parameters/PageSize"}
      responses:
        "200":
          description: Experiment page
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ExperimentPage"}
    post:
      tags: [Experiments]
      operationId: createExperiment
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ExperimentCreate"}
      responses:
        "202":
          description: Experiment accepted
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AcceptedExperiment"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /experiments/{experimentId}:
    parameters:
      - {$ref: "#/components/parameters/ExperimentId"}
    get:
      tags: [Experiments]
      operationId: getExperiment
      responses:
        "200":
          description: Experiment details
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Experiment"}
        "404": {$ref: "#/components/responses/Problem"}
  /experiments/{experimentId}/comparison:
    parameters:
      - {$ref: "#/components/parameters/ExperimentId"}
    get:
      tags: [Experiments]
      operationId: getExperimentComparison
      responses:
        "200":
          description: Comparison on identical rows
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ExperimentComparison"}
        "409": {$ref: "#/components/responses/Problem"}
  /experiments/{experimentId}/cancel:
    parameters:
      - {$ref: "#/components/parameters/ExperimentId"}
    post:
      tags: [Experiments]
      operationId: cancelExperiment
      responses:
        "200":
          description: Cancellation requested
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Experiment"}
        "409": {$ref: "#/components/responses/Problem"}
...
components:
  parameters:
...
    ExperimentId:
      name: experimentId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
  responses:
    Problem:
      description: Problem Details
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestId"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
  headers:
    RequestId:
      description: Correlation ID used in structured request logs.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        code: {type: string}
        request_id: {type: string}
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: {type: integer}
        page_size: {type: integer}
        total: {type: integer}
...
    ExperimentCreate:
      type: object
      required:
        [dataset_version_id, name, algorithms]
      properties:
        dataset_version_id: {type: string, format: uuid}
        name: {type: string, minLength: 1, maxLength: 160}
        weather_mode: {type: string, enum: [W0, W1]}
        algorithms:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum:
              [seasonal_naive_24, seasonal_naive_168, ridge,
               random_forest, hist_gradient_boosting]
        sensitivity_mode: {type: string, enum: [complete_only, coverage_90], default: complete_only}
    AcceptedExperiment:
      type: object
      required: [experiment_id, job_id, status]
      properties:
        experiment_id: {type: string, format: uuid}
        job_id: {type: string, format: uuid}
        status: {type: string, const: queued}
    MetricSet:
      type: object
      properties:
        mae: {type: number, minimum: 0}
        rmse: {type: number, minimum: 0}
        smape: {type: number, minimum: 0}
    ModelRun:
      type: object
      required: [id, algorithm, status, random_seed]
      properties:
        id: {type: string, format: uuid}
        algorithm: {type: string}
        status: {type: string}
        hyperparameters: {type: object}
        random_seed: {type: integer}
        cv_metrics: {$ref: "#/components/schemas/MetricSet"}
        final_test_metrics: {$ref: "#/components/schemas/MetricSet"}
        std_cv_mae: {type: number, minimum: 0}
        train_seconds: {type: number, minimum: 0}
        predict_ms_median: {type: number, minimum: 0}
        artifact_size_bytes: {type: integer, minimum: 0}
        is_recommended: {type: boolean}
    Experiment:
      type: object
      required:
        [id, dataset_version_id, job_id, name, status, weather_mode, sensitivity_mode,
         algorithms, created_at]
      properties:
        id: {type: string, format: uuid}
        dataset_version_id: {type: string, format: uuid}
        job_id: {type: string, format: uuid}
        name: {type: string}
        status: {type: string}
        weather_mode: {type: string, enum: [W0, W1]}
        sensitivity_mode: {type: string, enum: [complete_only, coverage_90]}
        algorithms:
          type: array
          items: {type: string}
        result_manifest: {type: object, nullable: true}
        failure_code: {type: string, nullable: true}
        failure_detail: {type: string, nullable: true}
        created_at: {type: string, format: date-time}
        started_at: {type: string, format: date-time}
        finished_at: {type: string, format: date-time}
    ExperimentPage:
      allOf:
        - {$ref: "#/components/schemas/PageMeta"}
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: {$ref: "#/components/schemas/Experiment"}
    ExperimentComparison:
      type: object
      required: [experiment_id, evaluation_rows, models]
      properties:
        experiment_id: {type: string, format: uuid}
        evaluation_rows: {type: integer}
        models:
          type: array
          items: {$ref: "#/components/schemas/ModelRun"}
        horizon_metrics:
          type: array
          items:
            type: object
            required: [model_run_id, horizon, mae, rmse, smape]
            properties:
              model_run_id: {type: string, format: uuid}
              horizon: {type: integer, minimum: 1, maximum: 24}
              mae: {type: number, minimum: 0}
              rmse: {type: number, minimum: 0}
              smape: {type: number, minimum: 0}
...